Stop clobbering view options in the clipboard plugin
I was switching to use Milkdown in one of my projects when I discovered
that the `ViewOptions` I was passing to `ProseMirror` were not being
respected.

E.g., I was passing in a `dispatchTransaction` prop:

```
ctx.set(
  editorViewOptionsCtx,
  {
    editable: () => !readOnly,
    dispatchTransaction(tr) {...}
  },
);
```

but it was never making it to the editor.

Putting a `console.log` in `core/internal-plugin/editor-view.ts` confirmed that the prop was
missing.

The culprit was the clipboard plugin -- it updates the view options
without preserving the existing options.